refactor: move completion helpers
