+++ internal/cmd/cmds/list.go
@@ -2,6 +2,7 @@ package cmds
 
 import (
 	"github.com/spf13/cobra"
+	"grog/internal/completions"
 	"grog/internal/config"
 	"grog/internal/label"
 	"grog/internal/loading"
@@ -22,7 +23,7 @@ var ListCmd = &cobra.Command{
   grog list //path/to/package/...       # List all targets in a package and subpackages
   grog list --target-type=test          # List only test targets`,
 	Args:              cobra.ArbitraryArgs, // Optional argument for target pattern
-	ValidArgsFunction: targetPatternCompletion,
+	ValidArgsFunction: completions.TargetPatternCompletion,
 	Run: func(cmd *cobra.Command, args []string) {
 		ctx, logger := setupCommand()
