+++ internal/cmd/cmds/test.go
@@ -2,6 +2,7 @@ package cmds
 
 import (
 	"github.com/spf13/cobra"
+	"grog/internal/completions"
 	"grog/internal/config"
 	"grog/internal/label"
 	"grog/internal/loading"
@@ -16,7 +17,7 @@ var TestCmd = &cobra.Command{
   grog test //path/to/package:test  # Run a specific test
   grog test //path/to/package/...   # Run all tests in a package and subpackages`,
 	Args:              cobra.ArbitraryArgs, // Optional argument for target pattern
-	ValidArgsFunction: targetPatternCompletion,
+	ValidArgsFunction: completions.TargetPatternCompletion,
 	Run: func(cmd *cobra.Command, args []string) {
 		ctx, logger := setupCommand()
